api/requirements.txt
pandas
numpy
faiss-cpu
psutil
wandb
h5py
torch-summary
flask
gunicorn
prometheus-client